Moved code to MissionTable and added comments
